+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 - Add -Raw parameter to Get-XMPlaylist and Get-XMStation
 - Refactor code
 - Update help
+
+1.3.8
+- Add -Filter parameter to Get-XMStation
+- Update ConvertFrom-APIStation function
+- Update README.md